+++ js/study.js
@@ -36,7 +36,7 @@ function renderStudyCard() {
   html += '</div>';
 
   /* Flashcard */
-  html += '<div class="fc-box" id="fc-box" onclick="flipStudyCard()">';
+  html += '<div class="fc-box" id="fc-box" role="button" tabindex="0" onclick="flipStudyCard()">';
   html += '<div class="fc-inner">';
   html += '<div class="fc-face fc-front">';
   html += '<div class="fc-front-label">ENGLISH</div>';